requirements.txt

streamlit==1.35.0
pandas==2.2.2

python-barcode==0.15.1

reportlab==4.1.0

Pillow==10.3.0

openpyxl==3.1.2
